+++ pyrtcm/rtcmhelpers.py
@@ -423,3 +423,15 @@ def id2prnsigmap(ident: str) -> tuple:
         SIGMAP = None
 
     return (PRNMAP, SIGMAP)
+
+
+def escapeall(val: bytes) -> str:
+    """
+    Escape all byte characters e.g. b'\\\\x73' rather than b`s`
+
+    :param bytes val: bytes
+    :return: string of escaped bytes
+    :rtype: str
+    """
+
+    return "b'{}'".format("".join(f"\\x{b:02x}" for b in val))